+++ Headers/gsc/GSGStateOps.h
@@ -33,16 +33,16 @@
 /* ----------------------------------------------------------------------- */
 /* Color operations */
 /* ----------------------------------------------------------------------- */
-- (void) DPScurrentalpha: (float*)a;
-- (void) DPScurrentcmykcolor: (float*)c : (float*)m : (float*)y : (float*)k;
-- (void) DPScurrentgray: (float*)gray;
-- (void) DPScurrenthsbcolor: (float*)h : (float*)s : (float*)b;
-- (void) DPScurrentrgbcolor: (float*)r : (float*)g : (float*)b;
-- (void) DPSsetalpha: (float)a;
-- (void) DPSsetcmykcolor: (float)c : (float)m : (float)y : (float)k;
-- (void) DPSsetgray: (float)gray;
-- (void) DPSsethsbcolor: (float)h : (float)s : (float)b;
-- (void) DPSsetrgbcolor: (float)r : (float)g : (float)b;
+- (void) DPScurrentalpha: (CGFloat*)a;
+- (void) DPScurrentcmykcolor: (CGFloat*)c : (CGFloat*)m : (CGFloat*)y : (CGFloat*)k;
+- (void) DPScurrentgray: (CGFloat*)gray;
+- (void) DPScurrenthsbcolor: (CGFloat*)h : (CGFloat*)s : (CGFloat*)b;
+- (void) DPScurrentrgbcolor: (CGFloat*)r : (CGFloat*)g : (CGFloat*)b;
+- (void) DPSsetalpha: (CGFloat)a;
+- (void) DPSsetcmykcolor: (CGFloat)c : (CGFloat)m : (CGFloat)y : (CGFloat)k;
+- (void) DPSsetgray: (CGFloat)gray;
+- (void) DPSsethsbcolor: (CGFloat)h : (CGFloat)s : (CGFloat)b;
+- (void) DPSsetrgbcolor: (CGFloat)r : (CGFloat)g : (CGFloat)b;
 
 - (void) GSSetFillColorspace: (void *)spaceref;
 - (void) GSSetStrokeColorspace: (void *)spaceref;
@@ -52,21 +52,21 @@
 /* ----------------------------------------------------------------------- */
 /* Text operations */
 /* ----------------------------------------------------------------------- */
-- (void) DPSashow: (float)x : (float)y : (const char*)s;
-- (void) DPSawidthshow: (float)cx : (float)cy : (int)c : (float)ax : (float)ay 
-		      : (const char*)s;
+- (void) DPSashow: (CGFloat)x : (CGFloat)y : (const char*)s;
+- (void) DPSawidthshow: (CGFloat)cx : (CGFloat)cy : (int)c
+                      : (CGFloat)ax : (CGFloat)ay : (const char*)s;
 - (void) DPScharpath: (const char*)s : (int)b;
 - (void) appendBezierPathWithPackedGlyphs: (const char *)packedGlyphs
                                      path: (NSBezierPath*)aPath;
 - (void) DPSshow: (const char*)s;
-- (void) DPSwidthshow: (float)x : (float)y : (int)c : (const char*)s;
-- (void) DPSxshow: (const char*)s : (const float*)numarray : (int)size;
-- (void) DPSxyshow: (const char*)s : (const float*)numarray : (int)size;
-- (void) DPSyshow: (const char*)s : (const float*)numarray : (int)size;
+- (void) DPSwidthshow: (CGFloat)x : (CGFloat)y : (int)c : (const char*)s;
+- (void) DPSxshow: (const char*)s : (const CGFloat*)numarray : (int)size;
+- (void) DPSxyshow: (const char*)s : (const CGFloat*)numarray : (int)size;
+- (void) DPSyshow: (const char*)s : (const CGFloat*)numarray : (int)size;
 
-- (void) GSSetCharacterSpacing: (float)extra;
+- (void) GSSetCharacterSpacing: (CGFloat)extra;
 - (void) GSSetFont: (GSFontInfo *)fontref;
-- (void) GSSetFontSize: (float)size;
+- (void) GSSetFontSize: (CGFloat)size;
 - (NSAffineTransform *) GSGetTextCTM;
 - (NSPoint) GSGetTextPosition;
 - (void) GSSetTextCTM: (NSAffineTransform *)ctm;
@@ -81,29 +81,29 @@
 /* ----------------------------------------------------------------------- */
 - (void) DPSinitgraphics;
 
-- (void) DPScurrentflat: (float*)flatness;
+- (void) DPScurrentflat: (CGFloat*)flatness;
 - (void) DPScurrentlinecap: (int*)linecap;
 - (void) DPScurrentlinejoin: (int*)linejoin;
-- (void) DPScurrentlinewidth: (float*)width;
-- (void) DPScurrentmiterlimit: (float*)limit;
-- (void) DPScurrentpoint: (float*)x : (float*)y;
+- (void) DPScurrentlinewidth: (CGFloat*)width;
+- (void) DPScurrentmiterlimit: (CGFloat*)limit;
+- (void) DPScurrentpoint: (CGFloat*)x : (CGFloat*)y;
 - (void) DPScurrentstrokeadjust: (int*)b;
-- (void) DPSsetdash: (const float*)pat : (int)size : (float)offset;
-- (void) DPSsetflat: (float)flatness;
+- (void) DPSsetdash: (const CGFloat*)pat : (NSInteger)size : (CGFloat)offset;
+- (void) DPSsetflat: (CGFloat)flatness;
 - (void) DPSsetlinecap: (int)linecap;
 - (void) DPSsetlinejoin: (int)linejoin;
-- (void) DPSsetlinewidth: (float)width;
-- (void) DPSsetmiterlimit: (float)limit;
+- (void) DPSsetlinewidth: (CGFloat)width;
+- (void) DPSsetmiterlimit: (CGFloat)limit;
 - (void) DPSsetstrokeadjust: (int)b;
 
 /* ----------------------------------------------------------------------- */
 /* Matrix operations */
 /* ----------------------------------------------------------------------- */
-- (void) DPSconcat: (const float*)m;
+- (void) DPSconcat: (const CGFloat*)m;
 - (void) DPSinitmatrix;
-- (void) DPSrotate: (float)angle;
-- (void) DPSscale: (float)x : (float)y;
-- (void) DPStranslate: (float)x : (float)y;
+- (void) DPSrotate: (CGFloat)angle;
+- (void) DPSscale: (CGFloat)x : (CGFloat)y;
+- (void) DPStranslate: (CGFloat)x : (CGFloat)y;
 
 - (NSAffineTransform *) GSCurrentCTM;
 - (void) GSSetCTM: (NSAffineTransform *)ctm;
@@ -114,32 +114,32 @@
 /* ----------------------------------------------------------------------- */
 - (NSPoint) currentPoint;
 
-- (void) DPSarc: (float)x : (float)y : (float)r : (float)angle1 
-	       : (float)angle2;
-- (void) DPSarcn: (float)x : (float)y : (float)r : (float)angle1 
-		: (float)angle2;
-- (void) DPSarct: (float)x1 : (float)y1 : (float)x2 : (float)y2 : (float)r;
+- (void) DPSarc: (CGFloat)x : (CGFloat)y : (CGFloat)r : (CGFloat)angle1 
+	       : (CGFloat)angle2;
+- (void) DPSarcn: (CGFloat)x : (CGFloat)y : (CGFloat)r : (CGFloat)angle1 
+		: (CGFloat)angle2;
+- (void) DPSarct: (CGFloat)x1 : (CGFloat)y1 : (CGFloat)x2 : (CGFloat)y2 : (CGFloat)r;
 - (void) DPSclip;
 - (void) DPSclosepath;
-- (void) DPScurveto: (float)x1 : (float)y1 : (float)x2 : (float)y2 
-		   : (float)x3 : (float)y3;
+- (void) DPScurveto: (CGFloat)x1 : (CGFloat)y1 : (CGFloat)x2 : (CGFloat)y2 
+		   : (CGFloat)x3 : (CGFloat)y3;
 - (void) DPSeoclip;
 - (void) DPSeofill;
 - (void) DPSfill;
 - (void) DPSflattenpath;
 - (void) DPSinitclip;
-- (void) DPSlineto: (float)x : (float)y;
-- (void) DPSmoveto: (float)x : (float)y;
+- (void) DPSlineto: (CGFloat)x : (CGFloat)y;
+- (void) DPSmoveto: (CGFloat)x : (CGFloat)y;
 - (void) DPSnewpath;
-- (void) DPSpathbbox: (float*)llx : (float*)lly : (float*)urx : (float*)ury;
-- (void) DPSrcurveto: (float)x1 : (float)y1 : (float)x2 : (float)y2 
-		    : (float)x3 : (float)y3;
-- (void) DPSrectclip: (float)x : (float)y : (float)w : (float)h;
-- (void) DPSrectfill: (float)x : (float)y : (float)w : (float)h;
-- (void) DPSrectstroke: (float)x : (float)y : (float)w : (float)h;
+- (void) DPSpathbbox: (CGFloat*)llx : (CGFloat*)lly : (CGFloat*)urx : (CGFloat*)ury;
+- (void) DPSrcurveto: (CGFloat)x1 : (CGFloat)y1 : (CGFloat)x2 : (CGFloat)y2 
+		    : (CGFloat)x3 : (CGFloat)y3;
+- (void) DPSrectclip: (CGFloat)x : (CGFloat)y : (CGFloat)w : (CGFloat)h;
+- (void) DPSrectfill: (CGFloat)x : (CGFloat)y : (CGFloat)w : (CGFloat)h;
+- (void) DPSrectstroke: (CGFloat)x : (CGFloat)y : (CGFloat)w : (CGFloat)h;
 - (void) DPSreversepath;
-- (void) DPSrlineto: (float)x : (float)y;
-- (void) DPSrmoveto: (float)x : (float)y;
+- (void) DPSrlineto: (CGFloat)x : (CGFloat)y;
+- (void) DPSrmoveto: (CGFloat)x : (CGFloat)y;
 - (void) DPSstroke;
 
 - (void) GSSendBezierPath: (NSBezierPath *)path;
@@ -149,9 +149,9 @@
 - (NSDictionary *) GSReadRect: (NSRect)rect;
 
 - (void)DPSimage: (NSAffineTransform*) matrix 
-		: (int) pixelsWide : (int) pixelsHigh
-		: (int) bitsPerSample : (int) samplesPerPixel 
-		: (int) bitsPerPixel : (int) bytesPerRow : (BOOL) isPlanar
+		: (NSInteger) pixelsWide : (NSInteger) pixelsHigh
+		: (NSInteger) bitsPerSample : (NSInteger) samplesPerPixel 
+		: (NSInteger) bitsPerPixel : (NSInteger) bytesPerRow : (BOOL) isPlanar
 		: (BOOL) hasAlpha : (NSString *) colorSpaceName
 		: (const unsigned char *const [5]) data;
