+++ Source/art/ftfont.h
@@ -36,7 +36,7 @@
 	color: (unsigned char)r : (unsigned char)g : (unsigned char)b
 	: (unsigned char)alpha
 	transform: (NSAffineTransform *)transform
-	deltas: (const float *)delta_data : (int)delta_size : (int)delta_flags
+	deltas: (const CGFloat *)delta_data : (int)delta_size : (int)delta_flags
         widthChar: (int) wch
 	drawinfo: (struct draw_info_s *)di;
 
@@ -60,7 +60,7 @@
 	drawinfo: (struct draw_info_s *)di;
 
 -(void) outlineString: (const char *)s
-	at: (float)x : (float)y
+	at: (CGFloat)x : (CGFloat)y
 	gstate: (void *)func_param;
 
 +(void) initializeBackend;
